Start implementing template class support
Add 'type-template' tag (and 'arg' child tag) to declare a definition of
a template struct/class. Modify 'object-type' and 'value-type' to allow
creating these as template instantiations referencing a 'type-template',
as an alternative to the usual non-template support.

There is also a 'redirect' attribute on 'arg' that will be used to
create "redirection" methods on the wrapped class to allow calling
methods of a template argument class as if they belonged to the
instantiation, which is sort of like having an operator->().

The eventual goal is to support wrapping of e.g. QSharedPointer.

This is sufficient to parse a typesystem XML (updates to 'sample'
included) using the new tags, but not to do anything useful yet, as
neither type resolution is working, nor do we generate code for the
instantiations.

Change-Id: I37960727c4dd69c39e84a414ae08752e3fd0dbf3