(Partly) implement wrappers for instantiations
Teach AbstractMetaBuilder to create wrappers for template instantiation
types. Teach generators to generate valid code for the same.

This is quite half-baked, as type resolution for instantiations is not
yet working (which tends to defeat the purpose). However, a typesystem
defining template instantiation types now generates code which also
compiles (and 'sample' is updated accordingly).

To facilitate code generation, fixedCppTypeName(QString) is now a public
static member of ShibokenGenerator, and is used in quite a few places
that previously only replaced '::'. This is needed as instantiation
wrappers have <>'s in their names that also need to be replaced, and
since the additional substitutions that fixedCppTypeName performs should
be harmless, this way we are using a helper function that already
existed rather than creating a new one for the purpose.

Change-Id: I7d077e89d9c3190e6850c1e04aac23a1f179c33b