Implement type look-up for template instantiatons
Teach translateType() to find and translate template instantiation
types. Teach buildAbstractMetaTypeFromString() how to create
AbstractMetaType's for the same.

This suffices to recognize and translate template instantiation types.
However, since redirection is not yet implemented, and adding methods
directly to the instantiation is not currently working, the wrapped
objects aren't of much use except as opaque handles.

However, this is enough to have at least the start of a meaningful unit
test, and one is accordingly added to 'sample'.

Change-Id: I3aee6ece8ad769a07127ff5fdd67a7af2d05200d