Remove useless code.
Remove code to set the type entry base container type in
AbstractMetaBuilder::setupInheritance outside of setting up the
container meta class. We do it anyway when the meta class is found, and
fail if we can't find the meta class, so having just the type entry set
is not really useful. (And this will allow us to simplify the code going
forward.)

Change-Id: I90389efc574d516703df0d6a26ba130769adfea7