Fix finding container base classes
Move logic to find template classes to its own helper method, and newly
use the helper in AbstractMetaBuilder::getBaseClasses. This fixes the
latter not finding template base classes. (However this is not
sufficient to get template base classes to appear in the correct
location in a topological sort, which is the eventual goal.)

Change-Id: I2a5e0df4035393288b6f69ebc41dc0e64866a182