Fix container argument resolution corner case
Fix a corner case where type resolution for a container type breaks due
to failing to add the template arguments, due to trailing whitespace
causing resolution of the same to fail. As best I can tell, this only
happens for redirector functions of type-template instantiations inside
of namespaces for which the type-template is also inside of a namespace.

Change-Id: Ie5934bcc03ada92c42df49e255ef6df743f9fac6