Fix resolution of instantiation arguments
Modify translateType to also try a canonicalized form of the type name.
This is necessary to resolve function arguments to type-template
instantiations, as the canonical instantiation name has spaces inside
the <>'s, but the function argument has these stripped due to minimal
signature normalization. (Return types are not normalized and so do not
have this problem.)

Change-Id: Ide191042c709f995a04ff7b77855c0e6f80a7d9c