Omit redirections for protected members
Don't generate redirections to internal class's protected members. This
isn't working right, because it requires a static cast of the inner
class to its wrapper to bypass the access protection, but right now the
generated code is doing the cast on the outer class instead (which is
otherwise the right thing to do). Since 1) it's questionable if
protected methods should be exposed anyway, and 2) likely significant
work to get it working, just disable it for now.

Change-Id: Ia7e3aee3ffd4f670fc74735cec3248dcf8f0d2ef