Always sanitize name of wrapper classes
It isn't obvious why only "inner classes" would need their names
sanitized; at least within our tests, no classes that trip this flag
have names that need to be sanitized. However, names of type template
instantiations must always be sanitized, and there is no ready check for
these at the moment, so just sanitize everything. (It may have just been
for performance reasons that we weren't doing this anyway, but this
shouldn't be in the critical path for that we need to be overly worried
about performance here.)

Change-Id: Ic13089efde9793fc5da8acd049e9d54363f5efa7