Use qualified name for wrapper class name
When generating wrapper classes, use the mangled qualified name instead
of just the base name. This is required to avoid symbol collisions in
the case of wrapping two classes with the same name in different
namespaces.

This doesn't "really" belong on this branch, as it applies to Shiboken
in general, however it effectively depends on the mangling that was
added here, and especially the previous commit, where we changed to
always mangle the wrapper name.

Change-Id: I276a97b4eac4385307a94159437ee4757d88a196