Fix type resolution for type-template functions
Modify type resolution of function modifications to accept a mapping of
template arguments, in order to support e.g. '%INTYPE_0' as return
and/or argument types to functions. This is relevant to function
modifications declared on type templates.

An example of such is added to the sample binding, however there is no
test yet, as functions added to instantiations via type templates are
not being added correctly at the moment. (This will be fixed in the next
commit.)

Change-Id: I8875354547aae2d56d66b5697f1672287717c7b0