Fix late resolution of instantiation arguments
Modify ShibokenGenerator::buildAbstractMetaTypeFromString to
canonicalize type names before trying to look up instantiations. Move
canonicalization function to AbstractMetaClass, and use when returning
the name of template instantiation classes.

This should ensure that everyone is looking at the same form of the type
name. This fixes a crash or assertion (depending on whether the type
appears by itself, or as a template type in a container-type) due to
failure to look up the type when it is used in code snippets.

Change-Id: Ic090138a9db9a3acbbd93a9015e07a9504080fd4