Add 'list output files' mode to shiboken
Split typesystem XML handler into itself and a new base class with some
common functionality. Create a new handler that guesses the names of
files that will be generated by shiboken and prints them. Add this as an
option to the shiboken executable.

This allows users to run shiboken to determine the list of generated
files, rather than having to specify them manually (which violates DRY).
This will be especially useful for integrating improved CMake support
(i.e. the ability to shiboken-wrap a library from CMake using only a
single CMake function, given an existing typesystem and list of
headers).

Change-Id: Ie6a0611dac1d2703d9b0af7cc54d10f2b0f794ad