Create 'use file' for Shiboken with CMake helpers
Create a 'use file' containing a CMake function 'sbk_wrap_library',
which leverages the new --list-outputs mode and some CMake magic to
greatly simplify wrapping libraries with Shiboken. This allows the
wrapping logic to be implemented as a single CMake call that handles
creating the rules to generate and build the wrapping code, as well as
some support for dependencies between bindings. It can also generate the
'global' header from a list of headers (e.g. projects that already have
a list of SDK headers to be installed), and in some (very simple) cases
can even create the entire typesystem XML.

Change-Id: Ie21e28a8b57bacb38ee10182086ccbf2f4427cbc