Document 'use file'
Add reStructuredText documentation to UseShiboken.cmake. This is done in
the upcoming CMake 3.0 style using lua-style bracketed comments, which
is convenient for editing as it is more likely that editors will be able
to apply syntax highlighting for this style. In order to maintain
compatibility with CMake 2.x, a small Python script converts these to
regular, single-line comments before the file is installed.

Change-Id: Ia4a963e5363bf865d31fad7cfe123a491efd5b8f