Provide a way to skip parsing statements.
Shiboken cannot parse many boost headers.
Shiboken will *never* be able to parse many boost headers.
When shiboken is preprocessing files, it should define a
macro so that it is possible to handle projects that include
boost without crashing.

This commit adds the macro SHIBOKEN_SKIP.
When it is defined, your source is being parsed by Shiboken.