More topology fixes
Remove concept of "injected dependencies"; this overloads the "injected"
term to mean something totally different, and no longer has any use
since the previous commit. Add a concept of "extra" dependencies, used
to form topology edges without actually modifying the class hierarchy,
which is used to ensure that pointer wrappers follow the classes that
they wrap. (This may not be strictly necessary, but it seems advisable.)
Add comments explaining better what traverseInstantiation is doing with
respect to the class hierarchy. Only treat a pointer-to-B as a subclass
of a pointer-to-A when there is only one template argument.