Fix photon test on Windows
MSVC 2015 is complaining about duplicate definitions between some
declared static variables in a template class in the header, and their
"this is where the symbol lives" declarations in the source file.

I'm not entirely sure why these were needed originally, and things seem
okay just removing the instances in the source file. At a guess, it's a
change since C++11. So long as there aren't link errors (and we're
generally requiring more modern compilers these days), I don't believe
the instances in the source file are needed.

Change-Id: Ibe563a0e3d93f66cde1422f8604b51484f5a77c9